Allow using a prestarted PoCL-D in the tests
If POCLD_PORT env is defined, the test scripts assume
a PoCL-D has been prespawned to that port and use it
instead.

Helps testing different PoCL-D configurations and
avoid the annoying port binding issues that seem
to pop up all the time.

+++ tools/scripts/test_remote_runner_with_local.sh
@@ -23,7 +23,12 @@ BUILD_DIR=$1
 TEST_BINARY=$2
 shift 2
 
-PORT=12000
+# If POCLD_PORT is defined, use a prelaunched PoCL-D at that port.
+if [ -z "$POCLD_PORT" ]; then
+ PORT=12000
+else
+ PORT=$POCLD_PORT
+fi
 
 echo "Running in $BUILD_DIR with PORT: $PORT"
 
@@ -42,10 +47,11 @@ export POCL_BUILDING=1
 export POCL_DEVICES="cpu"
 export POCL_DEBUG=
 
-$BUILD_DIR/pocld/pocld -a localhost -p $PORT -v error,warn,general &
-POCLD_PID=$!
-
-echo "Pocld running with PID: $POCLD_PID"
+if [ -z $POCLD_PORT ]; then
+  $BUILD_DIR/pocld/pocld -a localhost -p $PORT -v error,warn,general &
+  POCLD_PID=$!
+  echo "Pocld running with PID: $POCLD_PID"
+fi
 
 sleep 1
 
@@ -81,14 +87,19 @@ if [ -e "/proc/$EXAMPLE_PID" ]; then
   kill $EXAMPLE_PID
 fi
 
-if [ -e "/proc/$POCLD_PID" ]; then
-  kill $POCLD_PID
+if [ -z $POCLD_PORT ]; then
+  if [ -e "/proc/$POCLD_PID" ]; then
+    kill $POCLD_PID
+  fi
 fi
 
 sleep 2
 
 kill -9 $EXAMPLE_PID 1>/dev/null 2>&1
-kill -9 $POCLD_PID 1>/dev/null 2>&1
+
+if [ -z $POCLD_PORT ]; then
+    kill -9 $POCLD_PID 1>/dev/null 2>&1
+fi
 
 wait -f
